+++ src/services/auth.services.ts
@@ -1,8 +1,6 @@
 'use client'
-import { Account } from "@/Models/AccountModel";
+
 import axiosInstance from "./axios-instance";
-import { User } from "@/Models/UserModel";
-import { Bank } from "@/Models/BankModel";
 
 interface SignUpParams {
   email: string;
@@ -52,38 +50,3 @@ export const signOut = () => {
   window.location.href = "/sign-in";
 }
 
-export const getUser = async (): Promise<User> => {
-  return await axiosInstance()
-    .get("/users/")
-    .then((res) => {
-      return res.data;
-    })
-    .catch((error) => {
-      console.error(error);
-      throw error;
-    });
-}
-
-export const getAccounts = async (): Promise<Account[]> => {
-  return await axiosInstance()
-    .get("/accounts/")
-    .then((res) => {
-      return res.data;
-    })
-    .catch((error) => {
-      console.error(error);
-      throw error;
-    });
-}
-
-export const getBanks = async (): Promise<Bank[]> => { 
-  return await axiosInstance()
-    .get("/banks/")
-    .then((res) => {
-      return res.data;
-    })
-    .catch((error) => {
-      console.error(error);
-      throw error;
-    });
-}
