+++ src/services/user.services.ts
@@ -2,7 +2,7 @@ import { User } from "../Models/UserModel";
 import axiosInstance from "./axios-instance";
 export const getUser = async (): Promise<User> => {
   return await axiosInstance()
-    .get("/users/fetchUser")
+    .get("/users/")
     .then((res) => {
       return res.data;
     })
